Various changes
- Removed enable flag
- moved strings.py content to text/general.py
- renamed replace to replace_many
- key transformation enabled by default
