+++ constants/defaults.py
@@ -82,8 +82,7 @@ HTK_ENUM_SYMBOLIC_NAME_OVERRIDES = {}
 
 ##
 # JS Routes
-HTK_JS_ROUTES_ENABLED = False
-HTK_JS_ROUTER_USE_CAMEL_CASE_URL_NAMES = False
+HTK_JS_ROUTER_USE_CAMEL_CASE_URL_NAMES = True
 HTK_JS_ROUTES_URL_NAMES = []
 HTK_JS_ROUTES_DYNAMIC_PART_PLACEHOLDER = '{0}'
 HTK_JS_ROUTES_JS_OBJECT_NAME = 'window.Site'
