+++ templatetags/htk_tags.py
@@ -463,16 +463,13 @@ def credit_card_icon(credit_card_brand):
 @register.simple_tag()
 @mark_safe
 def print_js_routes():
-    if htk_setting('HTK_JS_ROUTES_ENABLED'):
-        from htk.utils.js_route_serializer import build_routes
-        js_obj = htk_setting('HTK_JS_ROUTES_JS_OBJECT_NAME')
-        urls = build_routes(as_json=True)
-        html = '<script type="text/javascript">'
-        html += '{js_obj} = Object.assign({js_obj} || {{}}, {{ urls: {urls} }});'.format(
-            js_obj=js_obj,
-            urls=urls
-        )
-        html += '</script>'
-    else:
-        html = ''
+    from htk.utils.js_route_serializer import build_routes
+    js_obj = htk_setting('HTK_JS_ROUTES_JS_OBJECT_NAME')
+    urls = build_routes(as_json=True)
+    html = '<script type="text/javascript">'
+    html += '{js_obj} = Object.assign({js_obj} || {{}}, {{ urls: {urls} }});'.format(
+        js_obj=js_obj,
+        urls=urls
+    )
+    html += '</script>'
     return html
